Kh-CMS-API/requirements.txt
django
django-rest-framework
django-cors-headers
mysqlclient
python-dotenv